Refactor Date class
Date is now shared by Task and SalesOrder objects. Have refactored
it to a more general package to reflect this.
