+++ src/main/java/seedu/address/logic/commands/EditTaskCommand.java
@@ -12,8 +12,8 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Date;
 import seedu.address.model.Model;
-import seedu.address.model.task.Date;
 import seedu.address.model.task.Label;
 import seedu.address.model.task.Task;
