+++ src/main/java/seedu/address/logic/parser/AddTaskCommandParser.java
@@ -8,7 +8,7 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.AddTaskCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.task.Date;
+import seedu.address.model.Date;
 import seedu.address.model.task.Label;
 import seedu.address.model.task.Task;
