+++ src/main/java/seedu/address/model/Date.java
@@ -1,4 +1,4 @@
-package seedu.address.model.task;
+package seedu.address.model;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
