+++ src/main/java/seedu/address/storage/JsonAdaptedTask.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.task.Date;
+import seedu.address.model.Date;
 import seedu.address.model.task.Label;
 import seedu.address.model.task.Task;
