+++ src/test/java/seedu/address/logic/commands/AddTaskCommandTest.java
@@ -10,7 +10,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.ModelStub;
-import seedu.address.model.task.Date;
+import seedu.address.model.Date;
 import seedu.address.model.task.Label;
 import seedu.address.model.task.Task;
