+++ src/test/java/seedu/address/model/salesorder/AmountTest.java
@@ -1,6 +1,8 @@
 package seedu.address.model.salesorder;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
