--- src/test/java/seedu/address/model/salesorder/SalesOrderTest.java
@@ -1,7 +1,0 @@
-package seedu.address.model.salesorder;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class SalesOrderTest {
-
-}
