+++ src/test/java/seedu/address/testutil/EditTaskDescriptorBuilder.java
@@ -2,7 +2,7 @@ package seedu.address.testutil;
 
 import seedu.address.logic.commands.EditTaskCommand;
 import seedu.address.logic.commands.EditTaskCommand.EditTaskDescriptor;
-import seedu.address.model.task.Date;
+import seedu.address.model.Date;
 import seedu.address.model.task.Label;
 import seedu.address.model.task.Task;
