+++ src/test/java/seedu/address/testutil/TaskBuilder.java
@@ -1,6 +1,6 @@
 package seedu.address.testutil;
 
-import seedu.address.model.task.Date;
+import seedu.address.model.Date;
 import seedu.address.model.task.Label;
 import seedu.address.model.task.Task;
